==> 4_compare_DE_DP.Rmd <==
---
title: "Plot results of disease progression"
author: Becker Kolja
date: 
output:
---

### Setup

```{r, setup, include=FALSE}

library(pcaMethods)
library(ggplot2)
library(cowplot)
library(RColorBrewer)
library(pheatmap)
library(reshape2)
library(data.table)
library(plyr)
library(dplyr)

source('./funcs.R')

# plotting options
colors.tissue = list(macula='#ed7d31', periphery='#5b9bd5')

```


## Load results from filtering and correction

```{r, load data}

# corrected data
iowa.full = fread('./data/human_retina_DR_totalRNA_normalized_cpm.txt') %>%
  data.frame %>%
  tibble::column_to_rownames('ensemblID')

# metdata
iowa.metadata = fread('./data/TableS1.csv') %>%
  dplyr::filter(sampleID %in% colnames(iowa.full)) %>%
  data.frame %>%
  tibble::column_to_rownames('sampleID')

# id conversion
df.map = fread('./data/IDconversion.txt')

```

# selected features from progression model

```{r, selected features}

# NOTE: this table contains beta values from final SPLS models
df.features.dp =
  fread('./results/df.features_progression_model.txt') %>%
  mutate(analysis = 'DP') %>%
  mutate(group = paste0(analysis, '.', sample_site)) %>%
  dplyr::select(ensemblID, analysis, sample_site, group)

```

# features identified from differential gene expression

```{r, }

df.features.de =
  fread('./results/df.features_differential_expression.txt') %>%
  dplyr::filter(disease_group == 'NPDR/PDR + DME') %>%
  dplyr::filter(adj.P.Val < 0.05) %>%
  mutate(analysis = 'DE') %>%
  mutate(group = paste0(analysis, '.', sample_site)) %>%
  dplyr::select(ensemblID, analysis, sample_site, group)

```

### comparison DE vs DP


```{r, overlap DP and DE genes}

# assemble all genes in one data table
df.genesDR.long = bind_rows(
  df.features.dp,
  df.features.de) %>%
  merge(df.map, by.x='ensemblID', by.y='OriginalName', all.x=TRUE) %>%
  add_count(ensemblID, name = 'score')

# save
fwrite(df.genesDR.long, file='./results/df.genesDR.long.txt')

# list of DR genes
groups = unique(df.genesDR.long$group)
genesDR = lapply(groups, function(x)
  df.genesDR.long[group == x, ensemblID]) %>%
  magrittr::set_names(groups)

```

```{r, upsetplot}

# comparison
compDR = compareSets(genesDR, rownames(iowa.full))
compLevels = c('DP Macula', 'DP Periphery', 'DE Macula', 'DE Periphery')

# ordering
groups.show =  rev(c(
  'DP Macula\nDE Macula', 'DP Periphery\nDE Periphery',
  'DE Macula\nDE Periphery', 'DP Macula\nDP Periphery', 
  'DP Macula\nDE Periphery', 'DP Periphery\nDE Macula'))

# nice labels
setTable = compDR$setTable %>%
  mutate(Var1 = gsub('[.]', ' ', Var1)) %>%
  mutate(Var2 = gsub('[.]', ' ', Var2)) %>%
  mutate(group = paste0(Var1, '\n', Var2)) %>%
  mutate(group = factor(group, levels=groups.show)) %>%
  data.table
setInfo = compDR$setInfo %>%
  mutate(setName = sub('[.]', ' ', setName)) %>%
  mutate(setName = factor(setName,
    levels = compLevels)) %>%
  data.table

# df for middle
df.cross = rbind(
  setTable[, c('Var1', 'group'), with=F], 
  setTable[, c('Var2', 'group'), with=F], 
  use.names=F) %>%
  mutate(Var1 = factor(Var1,
    levels = c('DP Macula', 'DP Periphery', 'DE Macula', 'DE Periphery'))) %>%
  mutate(group = factor(group,  levels = groups.show)) %>%
  data.table

# subplots
p.intersect = ggplot(setTable, aes(x=group, y=observed)) + 
  geom_bar(stat='identity', fill='gray66') +
  geom_text(aes(label=formatC(setTable$fdr, format='e', digits=0)),
    y=1, hjust=-0.1, color='black', size=1.5) +
  xlab(NULL) +
  ylab('intersect') +
  coord_flip() +
  theme(
    text=element_text(size=6),
    axis.text.y=element_blank(),
    axis.ticks=element_blank(),
    axis.title.x=element_text(hjust=0.5, vjust=20),
    plot.margin = unit(c(0, 0, 0, 0), "cm"))
p.sizes = ggplot(setInfo, aes(x=setName, y=setSize, fill=setName)) + 
  geom_bar(stat='identity') +
  scale_fill_manual(values =c('#f4a582', '#ca0020','#92c5de','#0571b0')) +
  labs(y='set size') +
  theme(
    text=element_text(size=6),
    axis.text.x=element_blank(),
    axis.ticks=element_blank(),
    axis.title.x=element_blank(),
    legend.position='none',
    plot.margin = unit(c(0, 0, 0, 0), "cm"))
p.cross = ggplot(df.cross, aes(y=group, x=Var1)) + 
  geom_point(size=0.5) +
  theme(
    text=element_text(size=6),
    plot.margin = unit(c(0, 0, 0, 0), "cm"),
    axis.title=element_blank(),
    axis.text.y=element_blank(),
    axis.text.x=element_text(angle=90, hjust=1, vjust=0.4),
    axis.ticks=element_blank())

# assemble
p.upset = p.sizes + plot_spacer() + p.cross + p.intersect + plot_layout(nrow=2)

# save
ggsave(plot=p.upset, filename='./figures/fig.upset.pdf',
       width=60, height=50, units='mm')

```
